[Refactor] Refactor model nparams calculation into single static function (#1767)
As titled, remove repeated copy
