+++ torchtitan/models/deepseek_v3/model/args.py
@@ -13,7 +13,8 @@ from torch import nn
 
 from torchtitan.config import JobConfig
 from torchtitan.models.moe import MoEArgs
-from torchtitan.protocols.train_spec import BaseModelArgs
+from torchtitan.models.utils import get_moe_model_nparams_and_flops
+from torchtitan.protocols.model import BaseModelArgs
 from torchtitan.tools.logging import logger
 from torchtitan.tools.utils import has_cuda_capability
 
@@ -105,57 +106,7 @@ class DeepSeekV3ModelArgs(BaseModelArgs):
                 "CP support for FlexAttention is still in progress."
             )
 
-    def get_nparams_and_flops(self, model: nn.Module, seq_len: int) -> tuple[int, int]:
-        """
-        Adopted from llama4 implementation.
-        """
-        nparams_embedding = 0
-        nparams_moe_router = 0
-        nparams_shared_experts = 0
-        nparams_experts = 0
-        nparams_dense = 0
-
-        for name, p in model.named_parameters():
-            if "embedding" in name:
-                nparams_embedding += p.numel()
-                nparams_dense += p.numel()
-            elif "moe.shared_experts" in name:
-                nparams_shared_experts += p.numel()
-            elif "moe.router" in name:
-                nparams_moe_router += p.numel()
-            elif "moe.experts" in name:
-                nparams_experts += p.numel()
-            else:
-                nparams_dense += p.numel()
-
-        nparams_sparse = nparams_moe_router + nparams_shared_experts + nparams_experts
-        nparams = nparams_dense + nparams_sparse
-        nparams_sparse_active = (
-            nparams_moe_router
-            + nparams_shared_experts
-            + nparams_experts * self.moe_args.top_k // self.moe_args.num_experts
-        )
-
-        logger.info(
-            f"Total parameter count: dense {nparams_dense:,}, "
-            f"sparse {nparams_sparse:,}, active {nparams_dense + nparams_sparse_active:,}"
-        )
-
-        l, h, q, t = (
-            self.n_layers,
-            self.n_heads,
-            self.dim // self.n_heads,
-            seq_len,
-        )
-        # Reasoning behind the factor of 12 for the self-attention part of the formula:
-        # 1. each self-attention has 2 matmul in the forward and 4 in the backward (6)
-        # 2. the flash attention does 1 more matmul recomputation in the backward
-        #    but recomputation should not be counted in calculating MFU           (+0)
-        # 3. each matmul performs 1 multiplication and 1 addition                 (*2)
-        # 4. we follow the convention and do not account for sparsity in causal attention
-        num_flops_per_token = (
-            6 * (nparams_dense - nparams_embedding + nparams_sparse_active)
-            + 12 * l * h * q * t
-        )
-
-        return nparams, num_flops_per_token
+    def get_nparams_and_flops(
+        self, model: nn.Module, seq_len: int
+    ) -> tuple[int, float]:
+        return get_moe_model_nparams_and_flops(self, model, seq_len)
