+++ torchtitan/models/llama3/model/args.py
@@ -12,7 +12,8 @@ from dataclasses import dataclass
 from torch import nn
 
 from torchtitan.config import JobConfig
-from torchtitan.protocols.train_spec import BaseModelArgs
+from torchtitan.models.utils import get_dense_model_nparams_and_flops
+from torchtitan.protocols.model import BaseModelArgs
 from torchtitan.tools.logging import logger
 
 
@@ -50,26 +51,7 @@ class TransformerModelArgs(BaseModelArgs):
                 "CP support for FlexAttention is still in progress."
             )
 
-    def get_nparams_and_flops(self, model: nn.Module, seq_len: int) -> tuple[int, int]:
-        nparams = sum(p.numel() for p in model.parameters())
-        nparams_embedding = sum(
-            sum(p.numel() for p in m.parameters())
-            for m in model.children()
-            if isinstance(m, nn.Embedding)
-        )
-
-        l, h, q, t = (
-            self.n_layers,
-            self.n_heads,
-            self.dim // self.n_heads,
-            seq_len,
-        )
-        # Reasoning behind the factor of 12 for the self-attention part of the formula:
-        # 1. each self-attention has 2 matmul in the forward and 4 in the backward (6)
-        # 2. the flash attention does 1 more matmul recomputation in the backward
-        #    but recomputation should not be counted in calculating MFU           (+0)
-        # 3. each matmul performs 1 multiplication and 1 addition                 (*2)
-        # 4. we follow the convention and do not account for sparsity in causal attention
-        num_flops_per_token = 6 * (nparams - nparams_embedding) + 12 * l * h * q * t
-
-        return nparams, num_flops_per_token
+    def get_nparams_and_flops(
+        self, model: nn.Module, seq_len: int
+    ) -> tuple[int, float]:
+        return get_dense_model_nparams_and_flops(self, model, seq_len)
